Updated networking to use bytebuf and handle packet headers more accurately.

+++ CenterServer/src/login/packet/LoopBackPacket.java
@@ -34,8 +34,8 @@ public enum LoopBackPacket {
         value = val;
     }
 
-    public int getValue() {
-        return value;
+    public short getValue() {
+        return (short) value;
     }
 
     public void setValue(int val) {
