+++ GameServer/src/center/packet/LoopBackPacket.java
@@ -30,8 +30,8 @@ public enum LoopBackPacket {
         value = val;
     }
 
-    public int getValue() {
-        return value;
+    public short getValue() {
+        return (short) value;
     }
 
     public void setValue(int val) {
