+++ GameServer/src/client/packet/LoopBackPacket.java
@@ -62,8 +62,8 @@ public enum LoopBackPacket {
         value = val;
     }
 
-    public int getValue() {
-        return value;
+    public short getValue() {
+        return (short) value;
     }
 
     public void setValue(int val) {
